# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ChecklistHQ API
  description: >-
    JSON API for ChecklistHQ (auth, checklists, notifications, users,
    templates, groups, tenant admin). Generated by hand from
    internal/api/*.go — see DESIGN.md for the broader system design.


    Auth model: cookie-based sessions (`checklists_session`, HttpOnly) set by
    login/register/password-reset-confirm. Every non-GET/HEAD request also
    requires a `X-CSRF-Token` header matching the (non-HttpOnly)
    `checklists_csrf` cookie value — a double-submit CSRF check — EXCEPT
    `/login`, `/register`, `/password-reset/request`, and
    `/password-reset/confirm`, which are unauthenticated by design and exempt.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local API (default api_port)
security:
  - sessionCookie: []
...
paths:
...
  /api/admin/users:
    get:
      tags: [Users]
      summary: "Admin: list users, sortable and filterable"
      description: >-
        Requires the caller to be an admin. Unlike GET /api/users (plain,
        unsorted, active users only), this endpoint supports sorting and can
        include inactive users. Mirrors the web UI's GET /admin/users/table.
      parameters:
        - name: sort
          in: query
          required: false
          schema: { type: string, enum: [name, username, email, is_admin, is_active] }
        - name: dir
          in: query
          required: false
          schema: { type: string, enum: [asc, desc] }
        - name: show_inactive
          in: query
          required: false
          description: "Pass 1 to include inactive users; omitted/any other value hides them."
          schema: { type: string, enum: ["1"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403": { description: Caller isn't an admin }
    post:
      tags: [Users]
      summary: "Admin: create a user"
      description: Requires the caller to be an admin.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { description: Missing required fields }
        "403": { description: Caller isn't an admin }
        "409": { description: Username already taken }
  /api/admin/users/{id}/active:
    post:
      tags: [Users]
      summary: "Admin: suspend or reactivate a user"
      description: >-
        Requires the caller to be an admin. Suspending a user also clears
        them as assignee/approver from any checklist in the same
        transaction. An admin cannot suspend their own account. Mirrors the
        web UI's POST /admin/users/{id}/active.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active: { type: boolean }
      responses:
        "204": { description: No Content }
        "400": { description: Invalid user id or malformed body }
        "403": { description: Caller isn't an admin, or is trying to suspend their own account }
  /api/admin/users/bulk:
    post:
      tags: [Users]
      summary: "Admin: bulk-create users from CSV"
      description: >-
        Raw CSV body, NO header row. Columns:
        username,password,name[,is_admin[,email]]. Requires admin. Rows are
        processed independently — a bad row doesn't abort the batch.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          text/csv:
            schema:
              type: string
              example: |
                carol,hunter22,Carol Example,false,carol@example.com
                dave,hunter22,Dave Example,true,dave@example.com
      responses:
        "200":
          description: Per-row results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BulkCreateUserResult" }
        "400": { description: Malformed CSV }
        "403": { description: Caller isn't an admin }
  /api/admin/users/export.csv:
    get:
      tags: [Users]
      summary: "Admin: export all users in the tenant as CSV"
      description: >-
        Requires admin. Columns: username,name,email,is_admin,is_active — the
        same columns the bulk-create endpoint accepts, minus password (a
        stored bcrypt hash can't be reversed into a usable password). Meant
        as a starting point for bulk edits (export, tweak in a spreadsheet,
        re-upload via /api/admin/users/bulk), not for recreating accounts
        verbatim. Mirrors the web UI's GET /admin/users/export.csv.
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                example: |
                  carol,Carol Example,carol@example.com,false,true
                  dave,Dave Example,dave@example.com,true,true
          headers:
            Content-Disposition:
              schema: { type: string, example: 'attachment; filename="users.csv"' }
        "403": { description: Caller isn't an admin }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: checklists_session
      description: >-
        Set by /login, /register, and /password-reset/confirm. Every
        non-safe-method request also needs a matching X-CSRF-Token header —
        see the CsrfToken parameter on each mutating operation.
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
      description: Must equal the value of the checklists_csrf cookie (double-submit CSRF check).
  schemas:
    User:
      type: object
      description: domain.User — field names are Go's default PascalCase JSON (no json tags on this struct, except PasswordHash which is never serialized).
      properties:
        ID: { type: integer, format: int64 }
        TenantID: { type: integer, format: int64 }
        Name: { type: string }
        Username: { type: string }
        Email: { type: string, nullable: true }
        IsAdmin: { type: boolean }
        IsActive: { type: boolean }
...
    CreateUserRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
        name: { type: string }
        email: { type: string }
        is_admin: { type: boolean }
      required: [username, password, name]
    BulkCreateUserResult:
      type: object
      properties:
        row: { type: integer }
        username: { type: string }
        status: { type: string, enum: [created, error] }
        error: { type: string }
        user: { $ref: "#/components/schemas/User" }
      required: [row, status]
...